fix: code issues
